# openapi/dist/openapi.json
{
  "components": {
    "schemas": {
...
      "ConfigValidateRequest": {
        "additionalProperties": false,
        "properties": {
          "translation_request": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TranslationRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "translator_config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TranslatorConfigValidateSpec"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "ConfigValidateRequest",
        "type": "object"
      },
      "ConfigValidateResponse": {
        "additionalProperties": false,
        "properties": {
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "schema_version": {
            "default": "1",
            "title": "Schema Version",
            "type": "string"
          },
          "translation_request_valid": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Translation Request Valid"
          },
          "translator_config_valid": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Translator Config Valid"
          }
        },
        "title": "ConfigValidateResponse",
        "type": "object"
      },
      "GlossaryEntrySpec": {
        "additionalProperties": false,
        "description": "Single glossary row (optional inline glossaries).",
        "properties": {
          "source": {
            "title": "Source",
            "type": "string"
          },
          "target": {
            "title": "Target",
            "type": "string"
          },
          "target_language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional tgt_lng column equivalent for filtering.",
            "title": "Target Language"
          }
        },
        "required": [
          "source",
          "target"
        ],
        "title": "GlossaryEntrySpec",
        "type": "object"
      },
      "GlossarySpec": {
        "additionalProperties": false,
        "description": "User glossary inputs: CSV paths and/or inline entries.",
        "properties": {
          "csv_paths": {
            "items": {
              "type": "string"
            },
            "title": "Csv Paths",
            "type": "array"
          },
          "inline_entries": {
            "items": {
              "$ref": "#/components/schemas/GlossaryEntrySpec"
            },
            "title": "Inline Entries",
            "type": "array"
          },
          "inline_name": {
            "default": "inline_glossary",
            "title": "Inline Name",
            "type": "string"
          }
        },
        "title": "GlossarySpec",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OpenAIRequestArgs": {
        "additionalProperties": false,
        "description": "Arguments for legacy OpenAI translator construction.",
        "properties": {
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "enable_json_mode_if_requested": {
            "default": false,
            "title": "Enable Json Mode If Requested",
            "type": "boolean"
          },
          "model": {
            "default": "gpt-4o-mini",
            "title": "Model",
            "type": "string"
          },
          "reasoning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning"
          },
          "send_dashscope_header": {
            "default": false,
            "title": "Send Dashscope Header",
            "type": "boolean"
          },
          "send_temperature": {
            "default": true,
            "title": "Send Temperature",
            "type": "boolean"
          },
          "term_api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Term Api Key"
          },
          "term_base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Term Base Url"
          },
          "term_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Term Model"
          },
          "term_reasoning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Term Reasoning"
          }
        },
        "title": "OpenAIRequestArgs",
        "type": "object"
      },
...
      "TranslationMemorySpec": {
        "additionalProperties": false,
        "description": "Translation memory options (maps to :class:`~doctranslate.format.pdf.translation_settings.TranslationSettings` TM fields).",
        "properties": {
          "tm_embedding_model": {
            "default": "sentence-transformers/all-MiniLM-L6-v2",
            "title": "Tm Embedding Model",
            "type": "string"
          },
          "tm_export_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tm Export Path"
          },
          "tm_fuzzy_min_score": {
            "default": 92.0,
            "title": "Tm Fuzzy Min Score",
            "type": "number"
          },
          "tm_import_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tm Import Path"
          },
          "tm_min_segment_chars": {
            "default": 12,
            "title": "Tm Min Segment Chars",
            "type": "integer"
          },
          "tm_mode": {
            "default": "off",
            "enum": [
              "off",
              "exact",
              "fuzzy",
              "semantic"
            ],
            "title": "Tm Mode",
            "type": "string"
          },
          "tm_project_id": {
            "default": "",
            "title": "Tm Project Id",
            "type": "string"
          },
          "tm_scope": {
            "default": "document",
            "enum": [
              "document",
              "project",
              "global"
            ],
            "title": "Tm Scope",
            "type": "string"
          },
          "tm_semantic_min_similarity": {
            "default": 0.9,
            "title": "Tm Semantic Min Similarity",
            "type": "number"
          }
        },
        "title": "TranslationMemorySpec",
        "type": "object"
      },
      "TranslationOptions": {
        "additionalProperties": false,
        "description": "PDF/runtime options (optional; defaults match engine defaults).",
        "properties": {
          "auto_extract_glossary": {
            "default": true,
            "title": "Auto Extract Glossary",
            "type": "boolean"
          },
          "debug": {
            "default": false,
            "title": "Debug",
            "type": "boolean"
          },
          "glossary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GlossarySpec"
              },
              {
                "type": "null"
              }
            ]
          },
          "llm_term_extraction_batch_max_paragraphs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Term Extraction Batch Max Paragraphs"
          },
          "llm_term_extraction_batch_max_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Term Extraction Batch Max Tokens"
          },
          "llm_translation_batch_max_paragraphs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Translation Batch Max Paragraphs"
          },
          "llm_translation_batch_max_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Translation Batch Max Tokens"
          },
          "max_pages_per_part": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Pages Per Part"
          },
          "metadata_extra_data": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata Extra Data"
          },
          "no_dual": {
            "default": false,
            "title": "No Dual",
            "type": "boolean"
          },
          "no_mono": {
            "default": false,
            "title": "No Mono",
            "type": "boolean"
          },
          "ocr_debug_dump": {
            "default": false,
            "title": "Ocr Debug Dump",
            "type": "boolean"
          },
          "ocr_lang_hints": {
            "items": {
              "type": "string"
            },
            "title": "Ocr Lang Hints",
            "type": "array"
          },
          "ocr_mode": {
            "default": "off",
            "enum": [
              "off",
              "auto",
              "force",
              "hybrid"
            ],
            "title": "Ocr Mode",
            "type": "string"
          },
          "ocr_pages": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ocr Pages"
          },
          "only_parse_generate_pdf": {
            "default": false,
            "title": "Only Parse Generate Pdf",
            "type": "boolean"
          },
          "output_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Dir"
          },
          "pages": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pages"
          },
          "pool_max_workers": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pool Max Workers"
          },
          "qps": {
            "default": 4,
            "title": "Qps",
            "type": "integer"
          },
          "report_interval": {
            "default": 0.1,
            "title": "Report Interval",
            "type": "number"
          },
          "skip_translation": {
            "default": false,
            "title": "Skip Translation",
            "type": "boolean"
          },
          "term_pool_max_workers": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Term Pool Max Workers"
          },
          "translation_memory": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TranslationMemorySpec"
              },
              {
                "type": "null"
              }
            ]
          },
          "use_rich_pbar": {
            "default": true,
            "title": "Use Rich Pbar",
            "type": "boolean"
          },
          "watermark_output_mode": {
            "default": "watermarked",
            "enum": [
              "watermarked",
              "no_watermark",
              "both"
            ],
            "title": "Watermark Output Mode",
            "type": "string"
          },
          "working_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Working Dir"
          }
        },
        "title": "TranslationOptions",
        "type": "object"
      },
      "TranslationRequest": {
        "additionalProperties": false,
        "description": "Stable, versioned request to run one PDF translation job.",
        "properties": {
          "input_pdf": {
            "description": "Path to the input PDF.",
            "title": "Input Pdf",
            "type": "string"
          },
          "lang_in": {
            "default": "en",
            "title": "Lang In",
            "type": "string"
          },
          "lang_out": {
            "default": "zh",
            "title": "Lang Out",
            "type": "string"
          },
          "options": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TranslationOptions"
              },
              {
                "type": "null"
              }
            ]
          },
          "schema_version": {
            "default": "1",
            "description": "Client-declared schema version; must match supported version.",
            "title": "Schema Version",
            "type": "string"
          },
          "translator": {
            "$ref": "#/components/schemas/TranslatorRequestConfig"
          }
        },
        "required": [
          "input_pdf"
        ],
        "title": "TranslationRequest",
        "type": "object"
      },
...
      "TranslatorConfigValidateSpec": {
        "additionalProperties": false,
        "description": "Optional nested validation for router/local TOML (mirrors CLI intent).",
        "properties": {
          "config_path": {
            "title": "Config Path",
            "type": "string"
          },
          "local_cli": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Local Cli"
          },
          "metrics_json_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics Json Path"
          },
          "metrics_output": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics Output"
          },
          "mode": {
            "enum": [
              "router",
              "local"
            ],
            "title": "Mode",
            "type": "string"
          },
          "routing_profile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Routing Profile"
          },
          "routing_strategy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Routing Strategy"
          },
          "term_extraction_profile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Term Extraction Profile"
          }
        },
        "required": [
          "mode",
          "config_path"
        ],
        "title": "TranslatorConfigValidateSpec",
        "type": "object"
      },
      "TranslatorMode": {
        "description": "How translators are constructed for a job.",
        "enum": [
          "openai",
          "router",
          "local"
        ],
        "title": "TranslatorMode",
        "type": "string"
      },
      "TranslatorRequestConfig": {
        "additionalProperties": false,
        "description": "How to build translators for a job.",
        "properties": {
          "cli_router_overrides": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cli Router Overrides"
          },
          "config_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Path"
          },
          "ignore_cache": {
            "default": false,
            "title": "Ignore Cache",
            "type": "boolean"
          },
          "local_cli": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Local Cli"
          },
          "mode": {
            "$ref": "#/components/schemas/TranslatorMode",
            "default": "openai"
          },
          "openai": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OpenAIRequestArgs"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "TranslatorRequestConfig",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "DocTranslater HTTP API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/config/validate": {
      "post": {
        "operationId": "v1_config_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigValidateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Validate Config",
        "tags": [
          "config"
        ]
      }
    },
...
  }
}